[FIX] l10n_cl_chart_of_account: AttributeError: 'account.invoice.tax' object has no attribute '_getNeto'
