+++ l10n_cl_chart_of_account/models/__init__.py
@@ -1,5 +1,8 @@
 # Copyright (c) 2018 Konos
 # Copyright (C) 2019 Open Source Integrators
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-from . import account_tax
-from . import account_invoice
+from . import (
+    account_tax,
+    account_invoice,
+    account_invoice_tax,
+)
